# Perf: short-circuit (+>) in ShortExpression mode