--- a/src/Fantomas/Context.fs
+++ b/src/Fantomas/Context.fs
@@ -308,7 +308,6 @@ let internal (+>) (ctx : Context -> Context) (f : _ -> Context) x =
     match y.WriterModel.Mode with
     | ShortExpression infos when infos |> Seq.exists (fun x -> x.ConfirmedMultiline) -> y
     | _ -> f y
-    //f (ctx x)
 
 /// Break-line and append specified string
 let internal (++) (ctx : Context -> Context) (str : string) x =