before

```
| Method |    Mean |   Error |  StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 13.35 s | 5.253 s | 0.288 s |    1 | 517000.0000 | 159000.0000 | 14000.0000 |   2.76 GB |
```

after

```
| Method |    Mean |   Error |  StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 11.02 s | 0.843 s | 0.046 s |    1 | 451000.0000 | 159000.0000 | 13000.0000 |   2.48 GB |
```